Move the resize logic to ByteStepperItemContent
